third_party/WebKit/LayoutTests/resize-observer/zoom.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<p>ResizeObserver page zoom test</p>
<div id="target1" style="width:100px;height:100px;">t1
</div>
<script>
'use strict';

let target = document.querySelector("#target1");

test( _ => {
    assert_false(window.internals == null,
      "test requires windows.internals.setZoomFactor");
    window.internals.setZoomFactor(1.5);
  },
  "windows.internals.setZoomFactor");

function test1() {
  let test = async_test("ResizeObserver zoomFactor");
  let width = 33;
  let ro = new ResizeObserver(entries => {
    test.step( _ => {
      assert_equals(entries[0].contentRect.width, width, "width is not zoomed");
      test.done();
      ro.disconnect();
    });
  });
  ro.observe(target);
  target.style.width = width + "px";
}
test1();

</script>
